Enable multiple selection in Playlist panel

+++ src/SoListBox.cpp
@@ -415,7 +415,7 @@ void guSoListBox::AppendFastEditMenu( wxMenu * menu, const int selcount ) const
 {
     wxMenuItem * MenuItem;
     //guLogMessage( wxT( "guSoLisBox::AppendFastEditMenu %i - %i" ), m_LastColumnRightClicked, m_LastRowRightClicked );
-    if( m_LastColumnRightClicked == wxNOT_FOUND || ( m_LastColumnRightClicked >= m_Columns->Count() ) )
+    if( m_LastColumnRightClicked == wxNOT_FOUND || ( m_LastColumnRightClicked >= ( int ) m_Columns->Count() ) )
     {
         return;
     }
